// src/api/generated/model/groupSparkyDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chargee Ampere
 * API v2
 * OpenAPI spec version: 2.0.0
 */

export interface GroupSparkyDto {
  /** The UUID of the sparky */
  uuid: string;
  /** The serial number of the sparky */
  serialNumber: string;
  /**
   * The box code of the sparky
   * @nullable
   */
  boxCode: string | null;
}